+++ tests/fuzz/fuzz.c
@@ -515,7 +515,29 @@ static int
     return 0;
 }
 
+static int fuzz_decode_pldm_rde_multipart_send_req(const struct pldm_msg* msg,
+                                                   size_t payload_length)
+{
+    struct pldm_rde_multipart_send_req req;
+
+    decode_pldm_rde_multipart_send_req(msg, payload_length, &req);
+
+    return 0;
+}
+
+static int fuzz_decode_pldm_rde_multipart_send_resp(const struct pldm_msg* msg,
+                                                    size_t payload_length)
+{
+    struct pldm_rde_multipart_send_resp resp;
+
+    decode_pldm_rde_multipart_send_resp(msg, payload_length, &resp);
+
+    return 0;
+}
+
 static int (*const decode_pldm_msg_tests[])(const struct pldm_msg*, size_t) = {
+    fuzz_decode_pldm_rde_multipart_send_req,
+    fuzz_decode_pldm_rde_multipart_send_resp,
     fuzz_decode_pldm_rde_operation_enumerate_req,
     fuzz_decode_pldm_rde_operation_enumerate_resp,
     fuzz_decode_pldm_rde_operation_status_req,