+++ CHANGELOG.md
@@ -19,6 +19,11 @@ Change categories:
 
 ### Added
 
+- rde: Add RDEMultipartSend command
+  - Add `encode_pldm_rde_rde_multipart_send_req()`
+  - Add `decode_pldm_rde_rde_multipart_send_req()`
+  - Add `encode_pldm_rde_rde_multipart_send_resp()`
+  - Add `decode_pldm_rde_rde_multipart_send_resp()`
 - rde: Add RDEOperationEnumerate command
   - Add `encode_pldm_rde_rde_operation_enumerate_req()`
   - Add `decode_pldm_rde_rde_operation_enumerate_req()`